requirements.txt
fastapi>=0.75.0
pynacl>=1.5.0
web3>=6.0.0
py-ipfs-http-client>=0.8.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
uvicorn>=0.17.0
python-dotenv>=0.19.0
solcx>=1.1.0
eth-account>=0.5.9
eth-utils>=2.1.0